dashboard/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
openai==1.12.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
streamlit>=1.28.0
plotly>=5.0.0
pandas>=2.0.0
beautifulsoup4>=4.12.0
supervisor>=4.2.0
pyodbc>=4.0.0
